Fix tests + finishes
